moving access code from applicant2Sequence
