+++ src/main/steps/applicant2/enter-your-access-code/get.test.ts
@@ -1,0 +1,40 @@
+import { defaultViewArgs } from '../../../../test/unit/utils/defaultViewArgs';
+import { mockRequest } from '../../../../test/unit/utils/mockRequest';
+import { mockResponse } from '../../../../test/unit/utils/mockResponse';
+import { DivorceOrDissolution } from '../../../app/case/definition';
+import { generatePageContent } from '../../common/common.content';
+
+import { generateContent } from './content';
+import { AccessCodeGetController } from './get';
+
+describe('PrivacyPolicyGetController', () => {
+  const controller = new AccessCodeGetController();
+  const language = 'en';
+
+  test('Should render the enter your access code page with divorce content', async () => {
+    const req = mockRequest();
+    const res = mockResponse();
+    await controller.get(req, res);
+    const isDivorce = true;
+
+    expect(res.render).toBeCalledWith(expect.anything(), {
+      ...generatePageContent({ language, pageContent: generateContent, isDivorce, userEmail: 'test@example.com' }),
+      ...defaultViewArgs,
+      formState: req.session.userCase,
+    });
+  });
+
+  test('Should render the enter your access code page with civil content', async () => {
+    const req = mockRequest();
+    const res = mockResponse();
+    res.locals.serviceType = DivorceOrDissolution.DISSOLUTION;
+    await controller.get(req, res);
+    const isDivorce = false;
+
+    expect(res.render).toBeCalledWith(expect.anything(), {
+      ...generatePageContent({ language, pageContent: generateContent, isDivorce, userEmail: 'test@example.com' }),
+      ...defaultViewArgs,
+      formState: req.session.userCase,
+    });
+  });
+});
