+++ src/main/steps/applicant2/enter-your-access-code/get.ts
@@ -1,0 +1,9 @@
+import { GetController } from '../../../app/controller/GetController';
+
+import { generateContent } from './content';
+
+export class AccessCodeGetController extends GetController {
+  constructor() {
+    super(__dirname + '/../../common/template.njk', generateContent);
+  }
+}
